remove pixel_renderer ok
